+++ ui/web_frontend/src/components/datatable/DataEditor.vue
