+++ ui/web_frontend/src/components/datatable/DataManager.vue
