+++ ui/web_frontend/src/components/datatree/DataList.vue
@@ -79,7 +79,7 @@ import { ref, onMounted, nextTick } from 'vue'
 import { ElMessage, } from 'element-plus'
 import { Folder, Document, Refresh } from '@element-plus/icons-vue'
 import { useI18n } from 'vue-i18n'
-import AddDialog from '@/components/manager/AddDialog.vue'
+import AddDialog from '@/components/datatable/AddDialog.vue'
 import AppNavbar from '@/components/support/AppNavbar.vue'
 import ContextMenu from './ContextMenu.vue'
 import { mapTreeData, updateNodeChildren, findNode } from './treeUtils'
