+++ ui/web_frontend/src/components/viewer/HighlightManager.js
