+++ ui/web_frontend/src/components/viewer/TextPlayer.vue
