+++ ui/web_frontend/src/components/viewer/TextPlayerManager.js
