+++ ui/web_frontend/src/components/viewer/ViewNote.vue
@@ -15,7 +15,7 @@ import { ref, onBeforeUnmount } from 'vue'
 import { useI18n } from 'vue-i18n'
 import { MdEditor } from 'md-editor-v3'
 import { ElMessage } from 'element-plus'
-import AddDialog from '@/components/manager/AddDialog.vue'
+import AddDialog from '@/components/datatable/AddDialog.vue'
 import { getDefaultPath, getDefaultVault } from './dataUtils';
 
 const emit = defineEmits(['noteChange'])
